# Conflicts:
#	.vscode/settings.json
#	pyproject.toml
#	remnapy/__init__.py
#	remnapy/controllers/hwid.py
#	remnapy/controllers/metadata.py
#	remnapy/controllers/node_plugins.py
#	remnapy/controllers/nodes.py
#	remnapy/controllers/subscription.py
#	remnapy/controllers/subscriptions_controller.py
#	remnapy/controllers/system.py
#	remnapy/controllers/users.py
#	remnapy/enums/__init__.py
#	remnapy/enums/auth.py
#	remnapy/models/__init__.py
#	remnapy/models/auth.py
#	remnapy/models/external_squads.py
#	remnapy/models/hosts.py
#	remnapy/models/infra_billing.py
#	remnapy/models/metadata.py
#	remnapy/models/node_plugins.py
#	remnapy/models/nodes.py
#	remnapy/models/passkeys.py
#	remnapy/models/remnawave_settings.py
#	remnapy/models/subscription_page.py
#	remnapy/models/system.py
#	remnapy/models/users.py
#	remnapy/models/users_bulk_actions.py
#	remnapy/models/webhook.py
#	test_imports.py
#	tests/test_bandwidthstats.py
#	tests/test_subscription.py

+++ remnapy/models/subscriptions_template.py
@@ -31,7 +31,7 @@ class GetTemplateResponseDto(TemplateResponseDto):
 
 
 class GetTemplatesData(BaseModel):
-    total: int
+    total: float
     templates: List[TemplateInfoDto]
