+++ remnapy/models/users_bulk_actions.py
@@ -137,7 +137,7 @@ class BulkAllExtendExpirationDateRequestDto(BaseModel):
 class BulkResponseData(BaseModel):
     """Common bulk response with affected rows"""
 
-    affected_rows: int = Field(alias="affectedRows")
+    affected_rows: float = Field(alias="affectedRows")
 
 
 class BulkEventResponseData(BaseModel):
